+++ R/theme_incompletes.R
@@ -9,6 +9,11 @@
 #'
 #' \tabular{ll}{
 #'    the_x45 \tab x-axis text in 45 angle \cr
+#'    the_legend_bot \tab legend.position = "bottom" \cr
 #'  }
 #'
 the_x45 <- function() theme(axis.text.x = element_text(angle = 45, hjust = 1))
+
+#' @describeIn the_x45
+#' @export
+the_legend_bot <- function() theme(legend.position = "bottom")