+++ NEWS.md
@@ -1,6 +1,7 @@
 ggptt 0.2.0
 ===========
-* `the_x45()` tab x-axis text in 45 angle.
+* `the_x45()` for x-axis text in 45 angle.
+* `the_legend_bot for legend.position = "bottom"`
 * `ptt_pal()` color PTT palette. 6 colours.
 * `scale_colour_ptt()` and `scale_fill_ptt()` for discrete PTT colour scales